refactor: page handler
