+++ packages/website/page.js
@@ -1,0 +1,5 @@
+/**
+ * @fileoverview Subpath export for the page handler factory.
+ */
+
+export { createPage } from '../../src/components/Page'
