+++ packages/website/reflections.js
@@ -3,4 +3,4 @@
  */
 
 export { default as reflectionStyles } from '../../src/components/Reflection.module.css'
-export { getEntries, getReflectionPageMap, getTags, reflectionsPageSize, Reflections, renderIndexPage, Tags, Title } from '../../src/components/Reflection'
+export { getEntries, getReflectionPageMap, getTags, reflectionsCollection, reflectionsPageSize, Reflections, renderIndexPage, Tags, Title } from '../../src/components/Reflection'
