+++ src/app/claude/[[...mdxPath]]/page.jsx
@@ -1,314 +1,20 @@
 /**
  * @fileoverview Dynamic page handler for the claude section.
  *
- * Serves all Nextra content pages through the catch-all route and
- * handles dynamic tag pages by intercepting reflections/tags/{tag}
- * paths. Reflection entries are served from R2 storage when available,
- * falling back to bundled MDX content. Tag pages render through the
- * same Wrapper component as MDX content, using MDX components for
- * consistent heading styling.
+ * Delegates to the shared subsite page factory bound to the claude
+ * subsite and reflections collection descriptors. Serves all Nextra
+ * content pages through the catch-all route and handles reflections
+ * entries, index pages, and tag pages via the collection descriptor.
  */
 
-import { Breadcrumb, PostCard, useMDXComponents as getMDXComponents } from '@axivo/website'
-import { reflections, subsite } from '@axivo/website/claude'
-import { domain } from '@axivo/website/global'
-import { getEntries, getTags, reflectionsPageSize, reflectionStyles, renderIndexPage } from '@axivo/website/reflections'
-import GithubSlugger from 'github-slugger'
-import { generateStaticParamsFor, importPage } from 'nextra/pages'
-import remarkMdx from 'remark-mdx'
-import remarkParse from 'remark-parse'
-import { SafeMdxRenderer } from 'safe-mdx'
-import { unified } from 'unified'
+import { subsite } from '@axivo/website/claude'
+import { createPage } from '@axivo/website/page'
+import { reflectionsCollection } from '@axivo/website/reflections'
 import '../../(home)/page.css'
 
-const components = getMDXComponents()
-const nextraStaticParams = generateStaticParamsFor('mdxPath')
-const reflectionsSection = reflections.section.slice(1)
-const Wrapper = components.wrapper
-
-/**
- * Builds an activePath array for Nextra's Breadcrumb component.
- * Each item needs route, name, title, and frontMatter to match
- * Nextra's Item type from normalize-pages.
- *
- * @param {string} tag - Tag name
- * @returns {object[]} Array of path items for Breadcrumb
- */
-function buildTagBreadcrumb(tag) {
-  return [
-    { name: reflectionsSection, route: reflections.path, title: reflections.title, frontMatter: {} },
-    { name: 'tags', route: `${reflections.path}/tags`, title: 'Tags', frontMatter: {} },
-    { name: tag, route: `${reflections.path}/tags/${tag}`, title: tag, frontMatter: {} }
-  ]
-}
-
-/**
- * Extracts table of contents from an MDAST tree.
- * Builds Nextra-compatible TOC entries (depth 2-6) with slugged IDs.
- * Annotates heading nodes with id properties for anchor link support.
- *
- * @param {object} mdast - MDAST root node
- * @returns {object[]} Array of { depth, value, id } heading entries
- */
-function extractToc(mdast) {
-  const slugger = new GithubSlugger()
-  const headings = []
-  for (const node of mdast.children) {
-    if (node.type === 'heading' && node.depth >= 2 && node.depth <= 6) {
-      const value = node.children.map(c => c.value || '').join('')
-      const id = slugger.slug(value)
-      node.data = { ...node.data, hProperties: { ...node.data?.hProperties, id } }
-      headings.push({ depth: node.depth, value, id })
-    }
-  }
-  return headings
-}
-
-/**
- * Fetches MDX content and metadata from R2 bucket.
- *
- * @param {string} key - R2 object key
- * @returns {Promise<{ content: string, metadata: object } | null>}
- */
-async function fetchR2Object(key) {
-  try {
-    const { getCloudflareContext } = await import('@opennextjs/cloudflare')
-    const { env } = await getCloudflareContext({ async: true })
-    const object = await env.CONTENT_BUCKET.get(key)
-    if (!object) {
-      console.warn(`R2 fetch returned null for key: ${key}`)
-      return null
-    }
-    const content = await object.text()
-    const metadata = { ...object.customMetadata }
-    if (metadata.description) {
-      metadata.description = decodeURIComponent(metadata.description)
-    }
-    if (metadata.tags) {
-      try {
-        metadata.tags = JSON.parse(metadata.tags)
-      } catch {
-        console.warn('Failed to parse tags metadata')
-      }
-    }
-    return { content, metadata }
-  } catch (error) {
-    console.error(`R2 fetch failed for key: ${key} - ${error.message}`)
-    return null
-  }
-}
-
-/**
- * Generates page metadata. Returns tag title for tag pages,
- * R2 metadata for reflection entries, or Nextra importPage metadata.
- *
- * @param {object} props - Next.js page props
- * @returns {Promise<object>} Page metadata with title
- */
-async function generateMetadata(props) {
-  const params = await props.params
-  const path = params.mdxPath || []
-  if (isReflectionEntry(path)) {
-    const key = `src/content/${subsite.path}/${path.join('/')}.mdx`
-    const result = await fetchR2Object(key)
-    if (result) {
-      return {
-        description: result.metadata.description,
-        title: result.metadata.title
-      }
-    }
-  }
-  if (isReflectionIndex(path)) {
-    const date = path.slice(1).join('/')
-    const { metadata } = await renderIndexPage(date)
-    return metadata
-  }
-  if (isTagPage(path)) {
-    return { title: decodeURIComponent(path[2]) }
-  }
-  const { metadata } = await importPage([subsite.path, ...path])
-  const result = { ...metadata }
-  if (result.seoTitle) {
-    result.title = result.seoTitle
-  }
-  if (!result.description) {
-    result.description = `${result.title} — ${subsite.title}`
-  }
-  return result
-}
-
-/**
- * Generates static params for all Nextra content pages and dynamic tag pages.
- *
- * @returns {Promise<object[]>} Array of path params for static generation
- */
-async function generateStaticParams() {
-  const params = await nextraStaticParams()
-  const sectionParams = params
-    .filter(p => p.mdxPath?.[0] === subsite.path)
-    .map(p => ({ mdxPath: p.mdxPath.slice(1) }))
-  const response = await fetch(`${domain.protocol}://${domain.name}/metadata`)
-  const { objects } = await response.json()
-  const indexDirs = new Set()
-  const reflectionParams = objects
-    .filter(obj => obj.template === 'blog')
-    .map(obj => {
-      const segments = obj.key
-        .replace('src/content/', '')
-        .replace('.mdx', '')
-        .split('/')
-        .slice(2)
-      const [year, month, day] = segments
-      indexDirs.add(`${year}`)
-      indexDirs.add(`${year}/${month}`)
-      indexDirs.add(`${year}/${month}/${day}`)
-      return { mdxPath: [reflectionsSection, ...segments] }
-    })
-  const indexParams = [...indexDirs].map(dir => ({
-    mdxPath: [reflectionsSection, ...dir.split('/')]
-  }))
-  const allTags = await getTags()
-  const tagParams = [...new Set(allTags)].map(tag => ({
-    mdxPath: [reflectionsSection, 'tags', tag]
-  }))
-  return [...sectionParams, ...reflectionParams, ...indexParams, ...tagParams]
-}
-
-/**
- * Checks if the path matches a reflection entry route.
- *
- * @param {string[]} path - URL path segments
- * @returns {boolean} True if path is reflections/YYYY/MM/DD/slug
- */
-function isReflectionEntry(path) {
-  return path[0] === reflectionsSection && path.length === 5 && /^\d{4}$/.test(path[1])
-}
-
-/**
- * Checks if the path matches a reflection index route.
- *
- * @param {string[]} path - URL path segments
- * @returns {boolean} True if path is reflections/YYYY, reflections/YYYY/MM, or reflections/YYYY/MM/DD
- */
-function isReflectionIndex(path) {
-  return path[0] === reflectionsSection && path.length >= 2 && path.length <= 4 && /^\d{4}$/.test(path[1])
-}
-
-/**
- * Checks if the path matches a tag page route.
- *
- * @param {string[]} path - URL path segments
- * @returns {boolean} True if path is reflections/tags/{tag}
- */
-function isTagPage(path) {
-  return path[0] === reflectionsSection && path[1] === 'tags' && path[2]
-}
-
-/**
- * Main page component for the claude section catch-all route.
- * Routes tag pages to renderTagPage, reflection entries to R2 content,
- * splash pages to full-width layout, and all other pages to the
- * standard Nextra docs wrapper.
- *
- * @param {object} props - Next.js page props
- * @returns {Promise<import('react').ReactElement>} Rendered page
- */
-async function Page(props) {
-  const params = await props.params
-  const path = params.mdxPath || []
-  if (isReflectionEntry(path)) {
-    const key = `src/content/${subsite.path}/${path.join('/')}.mdx`
-    const result = await fetchR2Object(key)
-    if (result) {
-      const mdast = parseMdx(result.content)
-      const r2Toc = extractToc(mdast)
-      return (
-        <Wrapper toc={r2Toc} metadata={result.metadata}>
-          <SafeMdxRenderer mdast={mdast} components={components} />
-        </Wrapper>
-      )
-    }
-  }
-  if (isReflectionIndex(path)) {
-    const date = path.slice(1).join('/')
-    const { content, metadata, toc } = await renderIndexPage(date)
-    return (
-      <Wrapper toc={toc} metadata={metadata}>
-        {content}
-      </Wrapper>
-    )
-  }
-  if (isTagPage(path)) {
-    return renderTagPage(decodeURIComponent(path[2]))
-  }
-  const pageModule = await importPage([subsite.path, ...path])
-  const {
-    default: MDXContent,
-    toc: originalToc,
-    metadata,
-    sourceCode
-  } = pageModule
-  const toc = [...originalToc]
-  const updateToc = (sectionId, items) => {
-    const index = toc.findIndex(item => item.id === sectionId)
-    if (index !== -1) {
-      toc.splice(index + 1, 0, ...items)
-    }
-  }
-  if (path.length === 1 && path[0] === reflectionsSection) {
-    const entries = await getEntries()
-    const latestToc = entries.slice(0, pageModule.reflectionsPageSize || reflectionsPageSize).map(entry => ({
-      depth: 3,
-      id: entry.route.split('/').pop(),
-      value: entry.frontMatter.title
-    }))
-    updateToc('latest-reflections', latestToc)
-  }
-  if (metadata.template === 'splash') {
-    return (
-      <div className="splash content-container">
-        <MDXContent {...props} params={params} />
-      </div>
-    )
-  }
-  return (
-    <Wrapper toc={toc} metadata={metadata} sourceCode={sourceCode}>
-      <MDXContent {...props} params={params} />
-    </Wrapper>
-  )
-}
-
-/**
- * Parses MDX content into an AST for safe rendering.
- *
- * @param {string} mdx - Raw MDX content
- * @returns {object} MDAST root node
- */
-function parseMdx(mdx) {
-  return unified().use(remarkParse).use(remarkMdx).parse(mdx)
-}
-
-/**
- * Renders a tag page with breadcrumb, heading, and filtered entries.
- *
- * @param {string} tag - Tag name
- * @returns {Promise<import('react').ReactElement>} Tag page content
- */
-async function renderTagPage(tag) {
-  const entries = await getEntries()
-  const filtered = entries.filter(entry =>
-    entry.frontMatter.tags?.includes(tag)
-  )
-  return (
-    <Wrapper toc={[]} metadata={{ title: tag }}>
-      <Breadcrumb activePath={buildTagBreadcrumb(tag)} />
-      <components.h1>{tag}</components.h1>
-      <components.h2 className={reflectionStyles.heading}>Reflections</components.h2>
-      {filtered.map(entry => (
-        <PostCard key={entry.route} post={entry} />
-      ))}
-    </Wrapper>
-  )
-}
+const { generateMetadata, generateStaticParams, Page } = createPage({
+  collection: reflectionsCollection,
+  subsite
+})
 
 export { generateMetadata, generateStaticParams, Page as default }
